UI branch to be merged with main (#1)
* Testin UI Branch

Added a UI Assets folder in order to test the UI Branch

* cleared

cleared the 3d objects from the scene

* Splash Screen created

Opened a new scene where I made the animations and buttons for the splash screen.

* Created Quit Popup with yes and no buttons inside. Made a button prefab.

* Created three additional Splash Screen buttons and popups for them.

* Splash Screen completed

* Started intro animation scene where the logos appear with fade in fade out.

* Intro screen image bugfix

* Quit Event Added

The "Yes" button found after clicking "Quit" now exits the game when clicked.

* Links

Added links that lead to the G.O website and the PNP discord to the corresponding UI buttons.

* Added Avatar Image for Pero and added intro animation with logos in a separate scene